Improvements on the application, to enable the Wing Scenario Attempt to get Armadillo up running again. Not working yet

+++ model/noc-types.h
@@ -57,7 +57,7 @@ enum SearchParam {
 
 typedef struct {
     int32_t x, y;
-    uint32_t value;
+    int64_t value;
     uint32_t serial_number; 
     bool updated;
 } NodeRef;
